+++ conf/default.conf.php
@@ -13,16 +13,21 @@ return static function (SeablastConfiguration $SBConfig): void {
         ->activate('as') // debug
         ->deactivate('mon'); // debug
     $SBConfig
+        // Debug
         ->setInt('a', 23)
         ->setInt('b', 45)
-        ->setInt(SeablastConstant::SB_ERROR_REPORTING, E_ALL & ~E_NOTICE)
         ->setString('test-string', 'default-value') // debug
         ->setArrayString('test-array-string', ['a', 'y', 'omega'])
+        // Environment
+        ->setInt(SeablastConstant::SB_ERROR_REPORTING, E_ALL & ~E_NOTICE)
         ->setInt(SeablastConstant::SB_INI_SET_SESSION_COOKIE_LIFETIME, 60 * 60 * 24 * 2) // 2 days
         ->setInt(SeablastConstant::SB_SESSION_SET_COOKIE_PARAMS_LIFETIME, 60 * 60 * 3) // 3 hours
         ->setString(SeablastConstant::SB_SESSION_SET_COOKIE_PARAMS_PATH, '/')
         ->setInt(SeablastConstant::SB_SETLOCALE_CATEGORY, LC_CTYPE)
         ->setString(SeablastConstant::SB_SETLOCALE_LOCALES, 'cs_CZ.UTF-8')
         ->setString(SeablastConstant::SB_ENCODING, 'UTF-8')
+        // Latte templates
+        ->setString(SeablastConstant::LATTE_TEMPLATE, 'templates')
+        ->setString(SeablastConstant::LATTE_CACHE, 'cache')
     ;
 };