+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - MVC architecture
 - SeablastConstant class for IDE hinting
+- Latte for rendering
 
 ### Changed